chore(build): package.sh call lint.sh & check_readme.sh before push

+++ bin/package.sh
@@ -37,4 +37,6 @@ echo -ne '#!python\n' | cat - $tmpzip > "$PVLIB_SH"
 chmod a+x "$PVLIB_SH"
 
 ## Nice chance to send tags.
-git push jrcstu latest --tag -f  # latest always forced.
+if { ./bin/check_readme.sh && ./bin/lint.sh; }; then
+    git push jrcstu latest --tag -f  # latest always forced.
+fi
